Move ssh session helper to api/sshutils.
