+++ api/observability/tracing/ssh/client_test.go
@@ -242,29 +242,3 @@ func TestSetEnvs(t *testing.T) {
 	default:
 	}
 }
-
-type mockSSHChannel struct {
-	ssh.Channel
-}
-
-func TestWrappedSSHConn(t *testing.T) {
-	sshCh := new(mockSSHChannel)
-	reqs := make(<-chan *ssh.Request)
-
-	// ensure that OpenChannel returns the same SSH channel and requests
-	// chan that wrappedSSHConn was given
-	wrappedConn := &wrappedSSHConn{
-		ch:   sshCh,
-		reqs: reqs,
-	}
-	retCh, retReqs, err := wrappedConn.OpenChannel("", nil)
-	require.NoError(t, err)
-	require.Equal(t, sshCh, retCh)
-	require.Equal(t, reqs, retReqs)
-
-	// ensure the wrapped SSH conn will panic if OpenChannel is called
-	// twice
-	require.Panics(t, func() {
-		wrappedConn.OpenChannel("", nil)
-	})
-}
